requirements.txt
openai-clip
opencv-python
Pillow
yaspin